Step2-2 기능 구현 완료 (#470)
* 2022-01-10 | YR | step3 까지 기능구현 완료

* 2022-01-10 | YR | pull kakao YoonRyeol

* 2022-01-13 | YR | 리뷰어 피드백 반영

* 2022-01-13 | YR | ExceptionController stackTrace 제거

* 2022-01-13 | YR | user update 예외처리

* 2022-01-13 | YR | user update 예외처리

* 2022-01-13 | YR | pull kakao YoonRyeol

* 2022-01-13 | YR | 접근제한자 수정, 폴더명 변경

* 2022-01-14 | YR | 리뷰어피드백 반영 : 예외처리 수정

* 2022-01-14 | YR | 리뷰어피드백 반영 : 예외처리 수정

* 2022-01-14 | YR | CustomException concrete 클래스로 수정

* 2022-01-14 | YR | 축약어 수정

* 2022-01-14 | YR | query > Parameterized query

* 2022-01-17 | YR | 2-1 구현완료

* 2022-01-18 | YR | 2-2 기능 구현완료

* 2022-01-18 | YR | DockerFile 추가

* 2022-01-18 | YR | pull kakao YoonRyeol

+++ src/main/java/com/kakao/cafe/util/exception/PostNotFoundException.java
@@ -1,7 +1,7 @@
 package com.kakao.cafe.util.exception;
 
 public class PostNotFoundException extends CustomException {
-    public PostNotFoundException(Exception e, int id) {
+    public PostNotFoundException(Exception e, long id) {
         super(e, String.format("Post에 해당 id가 존재하지 않습니다!", id));
     }
 }
